Update to 3.0.22

+++ PKGBUILD
@@ -21,8 +21,12 @@ build() {
 
 package() {
   cd "${srcdir}/CAD-${pkgver}"
-  install -Dm755 build/cad_desktop -t "${pkgdir}/usr/lib/hydracad"
-  install -Dm755 packaging/linux/cad_desktop.wrapper "${pkgdir}/usr/bin/cad_desktop"
+  if [ -f packaging/linux/cad_desktop.wrapper ]; then
+    install -Dm755 build/cad_desktop -t "${pkgdir}/usr/lib/hydracad"
+    install -Dm755 packaging/linux/cad_desktop.wrapper "${pkgdir}/usr/bin/cad_desktop"
+  else
+    install -Dm755 build/cad_desktop -t "${pkgdir}/usr/bin"
+  fi
   ln -s cad_desktop "${pkgdir}/usr/bin/hydracad"
   install -Dm644 packaging/linux/HydraCAD.desktop -t "${pkgdir}/usr/share/applications"
   install -Dm644 installer/license.txt "${pkgdir}/usr/share/licenses/${pkgname}/LICENSE"
